ExecutionInfo and ExecutedPayload Refactoring
