+++ crates/op-rbuilder/src/primitives/mod.rs
@@ -1,4 +1,7 @@
 //! This module contains types from the reth that weren't heavily modified
+mod execution;
 mod payload_builder_service;
 
 pub use payload_builder_service::PayloadBuilderService;
+
+pub use execution::{ExecutedPayload, ExecutionInfo};
